Bugfix: Stop audio from from being grabled static
By default `play-sound` will try `aplay` (from ALSA sound) if available.
`aplay` on my Ubuntu server in a VM  does not support mp3.
Maybe it's a codec issue, but it's easier to recommend another package.

I have added a `players` option to `play-sound`
that replicates the default but removes `aplay`